+++ sql/Pg/Bric/Util/Alerted.sql
@@ -12,10 +12,10 @@
 --
 
 CREATE TABLE alerted(
-    id           NUMERIC(10, 0)    NOT NULL
+    id           INTEGER           NOT NULL
                                    DEFAULT NEXTVAL('seq_alerted'),
-    usr__id      NUMERIC(10, 0)    NOT NULL,
-    alert__id    NUMERIC(10, 0)    NOT NULL,
+    usr__id      INTEGER           NOT NULL,
+    alert__id    INTEGER           NOT NULL,
     ack_time     TIMESTAMP,
     CONSTRAINT pk_alerted__id PRIMARY KEY (id)
 );
@@ -26,8 +26,8 @@ CREATE TABLE alerted(
 --
 
 CREATE TABLE alerted__contact_value(
-    alerted__id	            NUMERIC(10, 0)  NOT NULL,
-    contact__id             NUMERIC(10, 0)  NOT NULL,
+    alerted__id	            INTEGER         NOT NULL,
+    contact__id             INTEGER         NOT NULL,
     contact_value__value    VARCHAR(256)    NOT NULL,
     sent_time               TIMESTAMP,
     CONSTRAINT pk_alerted__contact_value PRIMARY KEY (alerted__id, contact__id, contact_value__value)
