+++ sql/Pg/Bric/Util/Grp/ContribType.sql
@@ -15,10 +15,10 @@ CREATE SEQUENCE seq_contrib_type_member START 1024;
 --
 
 CREATE TABLE contrib_type_member (
-    id          NUMERIC(10,0)  NOT NULL
+    id          INTEGER        NOT NULL
                                DEFAULT NEXTVAL('seq_contrib_type_member'),
-    object_id   NUMERIC(10,0)  NOT NULL,
-    member__id  NUMERIC(10,0)  NOT NULL,
+    object_id   INTEGER        NOT NULL,
+    member__id  INTEGER        NOT NULL,
     CONSTRAINT pk_contrib_type_member__id PRIMARY KEY (id)
 );
